split up databaseManager file
